changed API to get movies based on its genre ID bug fixed related to return of updated movie and genre in put call

+++ Server/models/movie.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.INTEGER
         },
         rating: {
-            type: DataTypes.TINYINT
+            type: DataTypes.FLOAT
         },
         imageURL: {
             type: DataTypes.STRING
